% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggTotals.R
\name{aggTotals}
\alias{aggTotals}
\title{Create EGSS-Table with calculated totals}
\usage{
aggTotals(x)
}
\arguments{
\item{x}{EGSS-data table with missing cells completed}
}
\value{
Completed EGSS data table with totals (total nace, total ceparemas
and totals) calculated. Observed totals
are replaced
}
\description{
Function caltulates all totals of EGSS data by summing up base-level data.
Function is applied directly after gap-filling.
No iterative fitting has to be applied after this step.
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016,
                 t1 = "TOT_EGSS")
datComp <- gapFill(x = datAll)
tabCalc <- aggTotals(datComp)

}
