% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency}
\alias{currency}
\title{Data Set with conversion rates for non-Eurozone countries}
\usage{
data(currency)
}
\description{
Conversion rates from national currency to Euro
\itemize{
\item time: Periods 2009 to 2018
\item rate: Conversion rates
\item geo: Country code: BG, CZ, DK, HR, HU, PL, RO, SE, UK
}
}
