% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EgssBas.R
\name{dat_egssBas}
\alias{dat_egssBas}
\title{EGSS Base Data Set}
\usage{
data(dat_egssBas)
}
\description{
EGSS-Base Data Set. Flat File with
\itemize{
\item Periods 1996 to 2016
\item all EU-Countries
\item four Variables: PROD, EMP, VA and EXP
\item 43 Nace-Levels:  A, B, C, C10-C12, C13-C15, C16-C18, C19, C20, C21, C22_C23, C24_C25, C26,
C27, C28, C29_C30, C31-C33, D, E, E36, E37, E38, E39, F, G, H, I, J,
K, L, M, M69_M70, M71, M72, M73-M75, N, O, P, Q, R, S, T, TOTAL, U
\item 29 Ceparema-Levels: CEPA1, CEPA112_122, CEPA2, CEPA3, CEPA4, CEPA5, CEPA6, CEPA7, CEPA7-9, CEPA8,
CEPA812, CEPA9, CREMA10, CREMA11, CREMA11A, CREMA11B, CREMA12, CREMA12_15_16,
CREMA13, CREMA13A, CREMA13B, CREMA13C, CREMA14, CREMA15, CREMA15A, CREMA16,
TOT_CEPA, TOT_CREMA, TOTAL
\item 8 Type-Levels: ANC, C_REP, ES_CS, MKT, NMKT, OWN_USE, REST, TOT_EGSS
}
}
