% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraPol.R
\name{extraPol}
\alias{extraPol}
\title{Extrapolate missing EGSS-data}
\usage{
extraPol(x)
}
\arguments{
\item{x}{EGSS-Data set created with Function loadNA()}
}
\value{
completed list element for aggregation list
}
\description{
If EGSS-data for a certain time series (geo x indic_pi x nace_r2
x ty) is available for at least one period data is extrapolated backwards
and forwards in order to generate a complete time series from 2010 to the actual
period. The Forecasting method applied here are exponential smoothing state space
models (ETS) from the forecast package.
}
\details{
If only one value is available for the  whole time series or if the variance of
the variable is equal to zero for the whole time series forecasts and backcasts
are set to this very value. In any other case extrapolation (and intrapolation)
is performed with ETS models from the forecast-package (see description in the
vignette).
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016, t1 = "TOT_EGSS")
testSer <- datAll[code=="AT_EMP_A00_00_000",]
extraPol(testSer)
}
