% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapFill.R
\name{gapFill}
\alias{gapFill}
\title{Completion of all gaps in the data matrix}
\usage{
gapFill(x)
}
\arguments{
\item{x}{EGSS-data set enriched with national account data (output of function loadNA)}
}
\value{
y Completed EGSS-data set (gaps filled with extrapolation and country-specific medians)
}
\description{
Gap filling is processed in two steps:
}
\details{
\enumerate{
\item Extrapolation and interpolation with time-series models,
\item Imputation of country-specific medians
}

Both attempts dont consider the original reported values (variable obs_value) for processing, but the ratios of obs_values
to the respective values from national accounts. For the extrapolation step these ratios are projected backwards and forwards
(and in-between available observations) and the estimated observation value is then generated by applying these ratios to
the respective value from national accounts. For the second attempt, which is only applied if no single observation is
available for the hierarchy level under consideration, relative values of EGSS to national accounts data are calculated for
all countries who delivered data for the relevant variable (production, employees, value added or exports). From these
ratios a median ratio is applied to the national account data of the country to impute.
ATTENTION: The number of countries contributing to this median could differ from year to year!
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
datEgssNA <- loadNA(x = natAcc, y = datEgss, toEst = 2016, t1 = "TOT_EGSS")
datEgssNA <- datEgssNA[!(geo \%in\% c("TR","CH")),]
datEgssNA <- datEgssNA[!(indic_pi == "EXP"),]
datComp <- gapFill(x = datEgssNA)

}
