% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genConv.R
\name{genConv}
\alias{genConv}
\title{Iterative fitting over all marginal tables of EGSS-data}
\usage{
genConv(x, n)
}
\arguments{
\item{x}{Gap-filled EGSS data set (non consistent data)}

\item{n}{Number of iterations}
}
\value{
EGSS-data set made consistent in all dimensions
}
\description{
Generation of final (consistent) EGSS data-matrices from gap-filled EGSS data.
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
datEgssNA <- loadNA(x = natAcc, y = datEgss, toEst = 2016)
datComp <- gapFill(x = datEgssNA)
genConv(x = datComp, n = 10)

}
