% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterControl.R
\name{iterControl}
\alias{iterControl}
\title{Evaluation of iterative Fitting}
\usage{
iterControl(x)
}
\arguments{
\item{x}{Completed EGSS-data table}
}
\value{
Ratios of row/column- wise differences greater than 0.5 and 0.02 in absolute value together with
the value, country, year and variable of the maximum deviation
}
\description{
Function to check, if the iterative fitting converges. This function checks the horizontal and vertical
consistency of the imputed data set.
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
datEgssNA <- loadNA(x = natAcc, y = datEgss, toEst = 2016, t1 = "TOT_EGSS")
datComp <- gapFill(x = datEgssNA)
resPrelim <- genConv(datComp, 50)
iterControl(resPrelim)

}
