% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterFit.R
\name{iterFit}
\alias{iterFit}
\title{Iterative fitting of extrapolated and imputed data to respective available aggregate numbers}
\usage{
iterFit(x, z, lc = c(8, 0, 0), nLevel = "TOTAL", cLevel = "TOTAL",
  nLevel1 = "TOTAL", cLevel1 = "TOTAL")
}
\arguments{
\item{x}{Complete EGSS-data set, i.e. data-set with no gaps. This data-set is generated with the
function gapFill.R}

\item{z}{NACE or ceparema for merging input data set with sub-aggregate data-set}

\item{lc}{Vector which allows different aggregation levels.

possible values:
\itemize{
\item c(6,0,0) - Fitting total Nace-levels to EGSS-total and fitting total ceparema-levels to EGSS-total
\item c(8,0,0) - (default) Fitting row-entries to total nace-letters
\item c(6,15,17) - Fitting column-entries to total ceparema-codes
}}

\item{nLevel}{Top-level for NACE}

\item{cLevel}{Top-level for Ceparema}

\item{nLevel1}{Bottom-level for NACE}

\item{cLevel1}{Bottom-level for Ceparema}
}
\value{
Complete and consistent EGSS-data set regarding the respective aggregation level.
}
\description{
Iterative fitting of extrapolated and imputed data to respective available aggregate numbers
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
datEgssNA <- loadNA(x = natAcc, y = datEgss, toEst = 2016, t1 = "TOT_EGSS")
datComp <- gapFill(x = datEgssNA)
agg1 <- iterFit(datComp,z=NULL,lc=c(6,0,0),nLevel1=LETTERS[1:21])
agg2a <- iterFit(agg1,z=NULL,lc=c(6,0,0),cLevel1=c("TOT_CEPA", "TOT_CREMA"))
agg2b <- iterFit(agg2a,z=c("ceparema","TOT_CEPA"),lc=c(6,0,0), cLevel=c("TOT_CEPA"),
                cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9"))
agg2c <- iterFit(agg2b,z=c("ceparema","TOT_CREMA"),lc=c(6,0,0), cLevel=c("TOT_CREMA"),
                 cLevel1=c("CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16"))
agg3 <- iterFit(agg2c,z = c("ceparema","TOTAL"), nLevel=LETTERS[1:21], nLevel1=LETTERS[1:21],
                cLevel1=c("TOT_CEPA", "TOT_CREMA"))
agg4 <- iterFit(agg3,z = c("nace_r2","TOTAL"),lc=c(6,15,17),nLevel1=LETTERS[1:21],
                cLevel=c("TOT_CEPA", "TOT_CREMA"), cLevel1=c("TOT_CEPA", "TOT_CREMA"))
agg5 <- iterFit(agg4,z=c("ceparema","TOT_CEPA"), nLevel=LETTERS[1:21],nLevel1=LETTERS[1:21],
                cLevel=c("TOT_CEPA"),
                cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9"))
agg6 <- iterFit(agg5,z=c("ceparema","TOT_CREMA"), nLevel=LETTERS[1:21],nLevel1=LETTERS[1:21],
                cLevel=c("TOT_CREMA"),
                cLevel1=c("CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16"))
agg7 <- iterFit(agg6,z=c("nace_r2","TOTAL"),lc=c(6,15,17),nLevel1=LETTERS[1:21],
         cLevel=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9","TOT_CEPA",
                  "CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16","TOT_CREMA"),
         cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9","TOT_CEPA",
                 "CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16","TOT_CREMA"))
}
