% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEGSS.R
\name{loadEGSS}
\alias{loadEGSS}
\title{Load and Compile EGSS-data}
\usage{
loadEGSS(x)
}
\arguments{
\item{x}{R Data Table (EGSS Basis File in Flat Format)}
}
\value{
Completed EGSS-Data Matrix
}
\description{
Loads EGSS Base File, completes and harmonises data file over periods.
}
\details{
\enumerate{
\item Input File has to be available in the "MDT"-flat format
\item Standardisation of "ty"-Variable:
\itemize{
\item create structure for "ty" according to requirements 2013 (levels: MKT, C_REP and ES_CS)
\item add a Rest Category: Rest=MKT-C_REP-ES_CS
\item transform data structure for periods before 2013 to this very structure, i.e.
+ C_REP=GDA+TCI (if both available "old" C_REP to delete)
+ Rest=GDC+TCE
\item Treatment of NAs (e.g. GDA=200, TCI=NA; Sum=200 or NA)
}
\item Transfer "Basic Structure" to all Countries and Years (34.916 rows)
\itemize{
\item 4 variables (Production,Value Added,Employment and Exports)
\item 43 nace-codes (Nace-2digit levels, Nace-divisions and Totals)
\item 29 Ceparema codes (with aggregates)
\item 8 ty-codes and 4 ty-codes respectively (indic_pi="EXP")
}
\item Cells with Zeros and obs_status="L" (not applicable) are set to NA
}
}
\examples{
loadEGSS(dat_egssBas)
}
