% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNA.R
\name{loadNA}
\alias{loadNA}
\title{Load and Compile National Accounts Data. Merge NA-Data with EGSS-data}
\usage{
loadNA(x, y, toEst, t1 = "TOT_EGSS")
}
\arguments{
\item{x}{National Accounts File for Variable Production, Exports, Value added and Hours worked as R data table}

\item{y}{EGSS-Data set created with Function loadEGSS()}

\item{toEst}{Year to be predicted}

\item{t1}{Type variable for which the analysis should be performed. Possible values are TOT_EGSS (default),
MKT, NMKT, ANC, C_REP, ES_CS, OWN_USE and REST}
}
\value{
Complete EGSS-Data Matrix enriched with National Account Data and respective ratios of
EGSS-Vaues to National Account Data
}
\description{
Loads Data from National Accounts for all 28 EU-Countries and the Variables Production, Value Added,
Exports and Hours worked (instead of Empoyees - no full-time equivalents are available). In a second step the
national account data set is merged with the EGSS-data set produced with loadEGSS. For all time series a relative observation
value is calculated relative to the "nearest" national account aggregate which is continuously greater than
zero for the whole time series.
}
\details{
An identification code is created in order to simlify the access to different aggregation levels. This code is a 17 digit
character variable composed of (Example: AT_EMP_A00_00_000) :
\enumerate{
\item two digits for country code followed by underscore (e.g. AT_,BE_,...)
\item three digits for variable (e.g. EMP_, OUT_,...)
\item three digits for NACE-level A00-U00: (e.g. A00_, E00_, E36_,...)
\item two digits for type:
}
\itemize{
\item 10_: anciliary
\item 20_: own use
\item 30_: non market
\item 40_: market
\item 41_: market ressource efficient
\item 42_: market environmental specific services
\item 43_: market rest category
}
\enumerate{
\item three digits for Ceparema code
}
\itemize{
\item 110-190: Environmental Protection Cepa 1-6
\item 200-260: Ressource Management CrEMa 10-16 -- with subcategories 111 and 112 (without 110) and 131,132,133 (without 130)
}
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
loadNA(x = natAcc, y = datEgss, toEst = 2016, t1 = "TOT_EGSS")
}
