% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natAccN.R
\name{natAccN}
\alias{natAccN}
\title{National Accounts Data Set}
\usage{
data(natAccN)
}
\description{
Data Set with National Account Information
\itemize{
\item Periods 2010 to 2018 (2010 to 2017 for Exports)
\item all 28 EU-Countries plus EU28 and EU27_2019
\item four Variables: PROD, EMP, VA and EXP
\item 41 Nace-Levels: TOTAL, A, B, C, C10-C12, C13-C15, C16-C18, C19, C20, C21, C22_C23, C24_C25,
C26, C27, C28, C29_C30, C31-C33, D, E, E36, E37-E39, F, G, H, I, J, K, L,
M, M69_M70, M71, M72, M73-M75, N, O, P, Q, R, S, T, U
}
}
