% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Visualisation of original together with imputed data for EGSS}
\usage{
plotData(x, y, geoC = "ALL", varC = "PROD", nC = "TOTAL", cC = "TOTAL")
}
\arguments{
\item{x}{EGSS-Base data set (ouput of loadNA.R)}

\item{y}{EGSS-Aggregate with calibrated values (output of iterFit.R)}

\item{geoC}{Country codes to display, default = ALL}

\item{varC}{Variable code, possible values are "PROD" (default), "EMP", "VA", "EXP"}

\item{nC}{Nace code, default = "TOTAL"}

\item{cC}{Ceparema code, default = "TOTAL"}
}
\value{
Time series plots of EGSS-data with different completion version
}
\description{
Plots for every hierarchy level desired are created with values from the x-data set together with those from the
y-data set. If the raw-data set (output of loadNA) is set as x-data then the plot shows which values had been
originally reported by the member country. These are then compared with the estimated values from the y-data set,
which can be any data set of raw-gap filled data or data which had been adapted by the iterative fitting procedure.
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016, t1 = "TOT_EGSS")
datCompl <- gapFill(x = datAll)
plotData(x = datAll, y = datCompl)
}
