% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenData.R
\name{screenData}
\alias{screenData}
\title{Screening of EGSS-data}
\usage{
screenData(
  x,
  g = c("AT", "BE", "BG", "CY", "CZ", "DE", "DK", "EE", "EL", "ES", "FI", "FR", "HR",
    "HU", "IE", "IT", "LT", "LU", "LV", "MT", " NL", "PL", "PT", "RO", "SE", "SI", "SK",
    "UK"),
  v = c("PROD", "VA", "EXP", "EMP"),
  per = "ALL",
  n2 = c("A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P",
    "Q", "R", "S", "T", "U"),
  cep = c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7", "CEPA8",
    "CEPA9", "CREMA10", "CREMA11", "CREMA12", "CREMA13A", "CREMA13B", "CREMA13C",
    "CREMA14", "CREMA15", "CREMA16"),
  t1 = c("ANC", "MKT", "NMKT", "OWN_USE"),
  p = 1,
  val = 1,
  ia = FALSE
)
}
\arguments{
\item{x}{EGSS-data table created from flat file and/or adapted with various gap-filling functions.}

\item{g}{Country code.}

\item{v}{Variable to plot. Possible values are PROD, VA, EXP and EMP.}

\item{per}{Year to be plotted (default per = "ALL").}

\item{n2}{Nace-code: Possible values A, B, C, C10-C12, C13-C15,C16-C18, C19, C20,
C21, C22_C23, C24_C25, C26,C27, C28, C29_C30, C31-C33, D, E, E36,
E37, E38,E39, F, G, H, I, J, K, L, M, M69_M70, M71,M72,
M73-M75, N, O, P, Q, R, S, T, U,TOTAL,}

\item{cep}{Ceparema-code. Possible values are: CEPA1, CEPA112_122, CEPA2, CEPA3,
CEPA4, CEPA5, CEPA6, CEPA7, CEPA7-9, CEPA8, CEPA812,
CEPA9, CREMA10, CREMA11, CREMA11A, CREMA11B, CREMA12,
CREMA12_15_16, CREMA13, CREMA13A, CREMA13B, CREMA13C,
CREMA14, CREMA15, CREMA15A, CREMA16, TOT_CEPA, TOT_CREMA,
TOTAL}

\item{t1}{Ty-variable to be displayed. Possible values are ANC, C_REP, ES_CS, MKT,
NMKT, OWN_USE, REST, TOT_EGSS}

\item{p}{Kind of plot. Possible values are 1 - Country x Year; 2 - Country x Nace; 3 - Country x Ceparema;
4 - Nace x Ceparema}

\item{val}{Visualisation of percentage of non-missing values (default - val=1) or percentage of values>0 (val=0)}

\item{ia}{Interactive plot (default ia=FALSE)}
}
\value{
Summery Tables, Graphs
}
\description{
Visualisation of the share of missing data in the EGSS data matrix
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
screenData(datEgss)
screenData(datEgss,per=2015,p=2)
screenData(datEgss,per=2015,p=3,t1="TOT_EGSS")
screenData(datEgss,per=2015,g="AT",p=4)
screenData(datEgss,per=2015,g=c("AT","BE"),p=4)

}
