% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables1.R
\name{tables1}
\alias{tables1}
\title{Tables of Nace times Ceparema for given countries, years and variables}
\usage{
tables1(x, per, g, v)
}
\arguments{
\item{x}{EGSS-data table}

\item{per}{Period demanded for the table}

\item{g}{Country to be displayed}

\item{v}{Variable to be displayed}
}
\description{
This function creates Tables with the same built as the original EGSS-data tables from Excel. Further, the tables
consider, if a cell entry was reported by the member-country or if it was estimated with the gap-filling procedure.
Finally, row-sums and column-sums are compared with respective row- and column-totals. These tables can be created
with every level of EGSS-data set, i.e. with original reported data or with some intermediate or final EGSS-data set.
}
\examples{
data("dat_egssBas")
data("natAcc")
datEgss <- loadEGSS(x = dat_egssBas)
datEgssNA <- loadNA(x = natAcc, y = datEgss, toEst = 2016, t1 = "TOT_EGSS")
datComp <- gapFill(x = datEgssNA)
agg1 <- iterFit(datComp,z=NULL,lc=c(6,0,0),nLevel1=LETTERS[1:21])
agg2a <- iterFit(agg1,z=NULL,lc=c(6,0,0),cLevel1=c("TOT_CEPA", "TOT_CREMA"))
agg2b <- iterFit(agg2a,z=c("ceparema","TOT_CEPA"),lc=c(6,0,0), cLevel=c("TOT_CEPA"),
              cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9"))
agg2c <- iterFit(agg2b,z=c("ceparema","TOT_CREMA"),lc=c(6,0,0), cLevel=c("TOT_CREMA"),
                 cLevel1=c("CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16"))
agg3 <- iterFit(agg2c,z = c("ceparema","TOTAL"), nLevel=LETTERS[1:21], nLevel1=LETTERS[1:21],
                cLevel1=c("TOT_CEPA", "TOT_CREMA"))
# fit column-total of tot_cepa and tot_crema to tot_cepa and tot_crema over all naces
agg4 <- iterFit(agg3,z = c("nace_r2","TOTAL"),lc=c(6,15,17),nLevel1=LETTERS[1:21],
                cLevel=c("TOT_CEPA", "TOT_CREMA"), cLevel1=c("TOT_CEPA", "TOT_CREMA"))
# fit row-totals of cepa1-cepa79 to row-total of tot_cepa
agg5 <- iterFit(agg4,z=c("ceparema","TOT_CEPA"), nLevel=LETTERS[1:21],nLevel1=LETTERS[1:21],
                cLevel=c("TOT_CEPA"),
                cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9"))
# fit row-totals of crema10-crema14 to row-total of tot_crema
agg6 <- iterFit(agg5,z=c("ceparema","TOT_CREMA"), nLevel=LETTERS[1:21],nLevel1=LETTERS[1:21],
              cLevel=c("TOT_CREMA"),
              cLevel1=c("CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16"))
# fit colum totals of ceparema to total ceparemas over all naces
agg7 <- iterFit(agg6,z=c("nace_r2","TOTAL"),lc=c(6,15,17),nLevel1=LETTERS[1:21],
        cLevel=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9","TOT_CEPA",
                 "CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16","TOT_CREMA"),
        cLevel1=c("CEPA1", "CEPA2", "CEPA3", "CEPA4", "CEPA5", "CEPA6", "CEPA7-9","TOT_CEPA",
                  "CREMA10", "CREMA11", "CREMA13", "CREMA14", "CREMA12_15_16","TOT_CREMA"))
tables1(agg7, per = 2015, g = "AT", v = "PROD")
}
