% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genConv.R
\name{genConv}
\alias{genConv}
\title{Iterative fitting over all marginal tables of EGSS-data}
\usage{
genConv(x, n)
}
\arguments{
\item{x}{Gap-filled EGSS data set (non consistent data)}

\item{n}{Number of iterations}
}
\value{
EGSS-data set made consistent in all dimensions
}
\description{
Generation of final (consistent) EGSS data-matrices from gap-filled EGSS data.
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016, t1 = "TOT_EGSS")
datComp <- gapFill(x = datAll)
genConv(x = datComp, n = 10)

}
