% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Putting together all results}
\usage{
results(x, y)
}
\arguments{
\item{x}{EGSS data matrix with some tables which did not converge in the iterative fitting procedure}

\item{y}{Non-consistent tables from x which were made consistent by relaxing the constraints}
}
\value{
list element with consistent EGSS-data for all countries and EU totals
}
\description{
Generation of final EGSS data-matrices from iterative fitted data.
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016, t1 = "TOT_EGSS")

}
