% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables1.R
\name{tables1}
\alias{tables1}
\title{Tables of Nace times Ceparema for given countries, years and variables}
\usage{
tables1(x, per, g, v)
}
\arguments{
\item{x}{EGSS-data table}

\item{per}{Period demanded for the table}

\item{g}{Country to be displayed}

\item{v}{Variable to be displayed}
}
\description{
This function creates Tables with the same built as the original EGSS-data tables from Excel. Further, the tables
consider, if a cell entry was reported by the member-country or if it was estimated with the gap-filling procedure.
Finally, row-sums and column-sums are compared with respective row- and column-totals. These tables can be created
with every level of EGSS-data set, i.e. with original reported data or with some intermediate or final EGSS-data set.
}
\examples{
datEgss <- loadEGSS(x = dat_egssBas, y = currency)
datAll <- loadNA(x = natAccN, y = datEgss, z = currency, toEst = 2016, t1 = "TOT_EGSS")
datComp <- gapFill(x = datAll)
resPrelim <- genConv(datComp, 20)
tables1(resPrelim, per = 2015, g = "AT", v = "PROD")
}
